+++ test/api_ss_protocol_test.rb
@@ -5,7 +5,7 @@ class ApiSSTest < Test::Unit::TestCase
     @http = mock
     @http.stubs(:active?).returns(true)
 
-    @api = MatrixSdk::Api.new 'https://example.com', protocols: :SS
+    @api = MatrixSdk::Api.new 'https://example.com', protocols: :SS, threadsafe: false
     @api.instance_variable_set :@http, @http
     @api.stubs(:print_http)
   end