+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## **Unreleased**
 
-- Adds some multi-thread usage support to the API (create your API/client with `threadsafe: :multithread`)
+- Adds some multi-thread usage support to the API (create your API/client with `threadsafe: :multithread/true/false`)
+    The API will currently default to running with multi-threaded requests. In case your application is single-threaded - or never performs requests from multiple threads - then you can set `threadsafe: true/false` to support connection reuse.
 - Changes room finding to ignore non-canonical aliases by default
 - Improves room alias handling